+++ newton/_src/solvers/mujoco/solver_mujoco.py
@@ -3552,11 +3552,11 @@ class SolverMuJoCo(SolverBase):
             dim=mj_data.naconmax,
             inputs=[
                 self.mjc_geom_to_newton_shape,
-                self.mjc_body_to_newton,
-                self.mjw_model.opt.cone == int(self._mujoco.mjtCone.mjCONE_PYRAMIDAL),
+                self.mjw_model.opt.cone,
                 mj_data.nacon,
                 mj_contact.pos,
                 mj_contact.frame,
+                mj_contact.friction,
                 mj_contact.dist,
                 mj_contact.dim,
                 mj_contact.geom,
@@ -3566,6 +3566,7 @@ class SolverMuJoCo(SolverBase):
                 self.mjw_model.geom_bodyid,
                 mj_data.xpos,
                 mj_data.xquat,
+                mj_data.njmax,
             ],
             outputs=[
                 contacts.rigid_contact_count,